Extracted common methods because circular import
